refactor(parser): extract variable substitution into substitution.ts
Move substituteAll, substituteVariables, dynamic variable resolution and
request variable resolution out of parser.ts into src/lib/substitution.ts.
parser.ts temporarily re-exports the moved API so existing imports keep
working; a follow-up updates importers and drops the re-exports.
